i#42: add User & Employee ManyToMany relation

+++ database/migrations/update/0_1/2026_02_27_115724_add_uuid_to_confidant_persons_table.php
@@ -14,7 +14,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('confidant_persons', function (Blueprint $table) {
-            $table->uuid()->after('id');
+            if (! Schema::hasColumn('confidant_persons', 'uuid')) {
+                $table->uuid()->after('id');
+            }
         });
     }
 
@@ -24,7 +26,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('confidant_persons', function (Blueprint $table) {
-            $table->dropColumn('uuid');
+            if (Schema::hasColumn('confidant_persons', 'uuid')) {
+                $table->dropColumn('uuid');
+            }
         });
     }
 };
